third_party/WebKit/LayoutTests/http/tests/misc/cross-origin-iframe-autofocus-denied-expected.txt

CONSOLE MESSAGE: line 16: PASS: The input element is not focused.
Focus should not be set to the input element of iframe, due to it has a different origin
